# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flower Shop API",
        "description": "API documentation for the Flower Shop system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main API server"
        }
    ],
    "paths": {
...
        "/api/import-receipts": {
            "get": {
                "tags": [
                    "Import Receipts"
                ],
                "summary": "Danh sách phiếu nhập",
                "operationId": "a1945e1b528a25e955d0e37c46848eb6",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ImportReceiptResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Import Receipts"
                ],
                "summary": "Tạo mới phiếu nhập",
                "operationId": "6b9a5a76d7eb7f5ca5695936fe675fa8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ImportReceiptStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportReceiptResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/import-receipts/{id}": {
            "get": {
                "tags": [
                    "Import Receipts"
                ],
                "summary": "Chi tiết phiếu nhập",
                "operationId": "6905b85316ab3b6d6596e3cc74c50f09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportReceiptResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Import Receipts"
                ],
                "summary": "Cập nhật phiếu nhập (ghi chú hoặc thay đổi toàn bộ chi tiết)",
                "operationId": "7ba69c4cc53c34c374b945829290a583",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ImportReceiptUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportReceiptResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Import Receipts"
                ],
                "summary": "Xóa phiếu nhập",
                "operationId": "f780d6b0a89c5b019a8d9689862b8f99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ImportReceiptStoreRequest": {
                "required": [
                    "import_date",
                    "details"
                ],
                "properties": {
                    "import_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-09"
                    },
                    "note": {
                        "type": "string",
                        "example": "Nhập lô hoa hồng"
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "flower_id": {
                                    "type": "integer",
                                    "example": 3
                                },
                                "quantity": {
                                    "type": "integer",
                                    "example": 100
                                },
                                "import_price": {
                                    "type": "number",
                                    "format": "double",
                                    "example": 9000
                                },
                                "status": {
                                    "type": "string",
                                    "example": "hoa tươi"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "ImportReceiptUpdateRequest": {
                "properties": {
                    "import_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-10"
                    },
                    "note": {
                        "type": "string",
                        "example": "Chỉnh sửa ghi chú"
                    },
                    "details": {
                        "description": "Nếu gửi, sẽ thay thế toàn bộ chi tiết bằng mảng mới",
                        "type": "array",
                        "items": {
                            "properties": {
                                "flower_id": {
                                    "type": "integer",
                                    "example": 3
                                },
                                "quantity": {
                                    "type": "integer",
                                    "example": 80
                                },
                                "import_price": {
                                    "type": "number",
                                    "format": "double",
                                    "example": 8800
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "ImportReceiptDetailResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 15
                    },
                    "flower_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "flower_name": {
                        "type": "string",
                        "example": "Hoa hồng đỏ"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 100
                    },
                    "used_quantity": {
                        "type": "integer",
                        "example": 20
                    },
                    "remaining_quantity": {
                        "type": "integer",
                        "example": 80
                    },
                    "import_price": {
                        "type": "number",
                        "format": "double",
                        "example": 9000
                    },
                    "import_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-09"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "double",
                        "example": 900000
                    }
                },
                "type": "object"
            },
            "ImportReceiptResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 9
                    },
                    "import_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-09"
                    },
                    "note": {
                        "type": "string",
                        "example": "Nhập lô hoa hồng"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "double",
                        "example": 900000
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ImportReceiptDetailResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập token JWT vào ô Authorization",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}